chatbot/requirements.txt
chromadb
huggingface-hub==0.25.1
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.9
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.131
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
streamlit~=1.32.0
python-dotenv~=0.21.0
pandas~=2.2.2
beautifulsoup4
telebot
langchain_experimental
pymongo
